Authorization is optional now

+++ app/src/main/java/com/disgust/sereda/recipe/screens/info/RecipeInfoViewModel.kt
@@ -1,5 +1,8 @@
 package com.disgust.sereda.recipe.screens.info
 
+import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.ui.ExperimentalComposeUiApi
 import com.disgust.sereda.recipe.data.RecipeRepository
 import com.disgust.sereda.recipe.screens.info.interaction.RecipeInfoState
 import com.disgust.sereda.recipe.screens.info.interaction.RecipeInfoUIEvent
@@ -7,12 +10,17 @@ import com.disgust.sereda.recipe.screens.info.model.RecipeInfo
 import com.disgust.sereda.utils.base.NavigatorViewModel
 import com.disgust.sereda.utils.base.UIEventHandler
 import com.disgust.sereda.utils.commonModel.RecipeFavoriteState
+import com.disgust.sereda.utils.commonModel.UserNotAuthDialogState
 import com.disgust.sereda.utils.doSingleRequest
+import com.disgust.sereda.utils.navigation.Screen
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
 
+@ExperimentalComposeUiApi
+@ExperimentalAnimationApi
+@ExperimentalMaterialApi
 @HiltViewModel
 class RecipeInfoViewModel @Inject constructor(
     private val repository: RecipeRepository
@@ -22,19 +30,33 @@ class RecipeInfoViewModel @Inject constructor(
         MutableStateFlow<RecipeInfoState>(RecipeInfoState.Loading)
     val recipeInfoState = _recipeInfoState.asStateFlow()
 
+    private val _userNotAuthDialogState = MutableStateFlow(UserNotAuthDialogState.HIDDEN)
+    val userNotAuthDialogState = _userNotAuthDialogState.asStateFlow()
+
     override fun onUIEvent(event: RecipeInfoUIEvent) {
         when (event) {
             is RecipeInfoUIEvent.StartScreen -> getRecipeInfo(event.id, event.state)
             is RecipeInfoUIEvent.ButtonRestartClick -> getRecipeInfo(event.id)
             is RecipeInfoUIEvent.ButtonAddToFavoriteClick -> {
-                _recipeInfoState.value.doAsStateIfPossible<RecipeInfoState.Success> { state ->
-                    val recipe = state.data
-                    if (recipe.favoriteState == RecipeFavoriteState.FAVORITE) {
-                        deleteRecipeFromFavorite(recipe)
-                    } else if (recipe.favoriteState == RecipeFavoriteState.NOT_FAVORITE) {
-                        addRecipeToFavorite(recipe)
+                if (isAuth()) {
+                    _recipeInfoState.value.doAsStateIfPossible<RecipeInfoState.Success> { state ->
+                        val recipe = state.data
+                        if (recipe.favoriteState == RecipeFavoriteState.FAVORITE) {
+                            deleteRecipeFromFavorite(recipe)
+                        } else if (recipe.favoriteState == RecipeFavoriteState.NOT_FAVORITE) {
+                            addRecipeToFavorite(recipe)
+                        }
                     }
+                } else {
+                    _userNotAuthDialogState.value = UserNotAuthDialogState.SHOWN
                 }
+            }
+            is RecipeInfoUIEvent.UserNotAuthDialogDismiss -> {
+                _userNotAuthDialogState.value = UserNotAuthDialogState.HIDDEN
+            }
+            is RecipeInfoUIEvent.UserNotAuthDialogConfirmButtonClick -> {
+                _userNotAuthDialogState.value = UserNotAuthDialogState.HIDDEN
+                navigate(Screen.GoogleAuth.route)
             }
         }
     }
@@ -80,4 +102,6 @@ class RecipeInfoViewModel @Inject constructor(
             )
         }
     }
+
+    private fun isAuth() = repository.isAuth()
 }
